requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
line-bot-sdk==3.11.0
pytesseract==0.3.10
Pillow==10.3.0
python-dotenv==1.0.1
